chore: moved sync libs to utils/sync
- prepare to make utils/async for Textual port
